Gotta make the packer/unpacker now

+++ crates/dev_game/resources/resources/defaults/shaders/voxel_terrain/data.func.glsl
@@ -1,6 +1,6 @@
 // Some voxel data that is generated in the first pass of the compute shader
 struct VoxelData {
-    int density;
+    float density;
     int biomeID;
     int materialID;
 };
